Merged PR 2268: OCCP-1825 Added AEM Regression Scripts to AEM Regression Suite
![image.png](https://dev.azure.com/acn-ctc/2c1b6c71-4ad6-4658-ad0c-f9ca6da4cfdd/_apis/git/repositories/80b64a5e-122a-4125-9f15-ee6b4bfd32b2/pullRequests/2268/attachments/image.png) TC-1800	Verify feaured list component is configurable on an authored page
TC-1801	Verify the display of error message when mandatory fields are not added while authoring the content
TC-1802	Verify the display of error message when min. threshold for the tile are not met when authroing the Informational featured list component while authoring the content
TC-1811	Verify the entire FL card is clickable when only URL is authored without Learn More on WideScreen
TC-1837	Verify feaured list component is configurable on an authored page
TC-1838	Verify the display of error message when mandatory fields are not added while authoring the content
TC-1839	Verify the display of error message when min. threshold for the tile are not met when authroing the featured list component while authoring the content